fix: enable auditing, ddl validation, and migration
