+++ server/src/main/kotlin/org/elaastix/server/core/infrastructure/config/SecurityConfiguration.kt
@@ -21,17 +21,14 @@ package org.elaastix.server.core.infrastructure.config
 
 import org.elaastix.server.authn.ElaastixAuthenticationFilter
 import org.elaastix.server.authn.ElaastixAuthenticationProvider
-import org.elaastix.server.users.entities.UserEntity
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
-import org.springframework.data.domain.AuditorAware
 import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration
 import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity
 import org.springframework.security.config.annotation.web.builders.HttpSecurity
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity
 import org.springframework.security.config.annotation.web.invoke
 import org.springframework.security.config.http.SessionCreationPolicy
-import org.springframework.security.core.context.SecurityContextHolder
 import org.springframework.security.data.repository.query.SecurityEvaluationContextExtension
 import org.springframework.security.web.SecurityFilterChain
 import org.springframework.security.web.authentication.AnonymousAuthenticationFilter
@@ -39,7 +36,6 @@ import org.springframework.security.web.util.matcher.AnyRequestMatcher
 import org.springframework.web.servlet.HandlerExceptionResolver
 import org.springframework.web.servlet.config.annotation.CorsRegistry
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer
-import java.util.Optional
 
 /**
  * Configuration for Spring Security features.
@@ -50,19 +46,7 @@ import java.util.Optional
 @Configuration
 @EnableWebSecurity
 @EnableMethodSecurity
-class SecurityConfiguration :
-	WebMvcConfigurer,
-	AuditorAware<UserEntity> {
-	/**
-	 * See https://docs.spring.io/spring-data/jpa/reference/auditing.html#auditing.auditor-aware
-	 */
-	override fun getCurrentAuditor(): Optional<UserEntity> =
-		Optional.ofNullable(
-			SecurityContextHolder.getContext().authentication
-				?.takeIf { it.isAuthenticated }
-				?.let { it.principal as? UserEntity? },
-		)
-
+class SecurityConfiguration : WebMvcConfigurer {
 	override fun addCorsMappings(registry: CorsRegistry) {
 		registry.addMapping("/openapi.json")
 	}
